// projects/shared/src/lib/enums/council-type.ts
export enum CouncilType {
    MEDICINE = 'MEDICINE',
    PSYCHOLOGY = 'PSYCHOLOGY',
    PHYSIOTHERAPY = 'PHYSIOTHERAPY',
    DENTISTRY = 'DENTISTRY',
    NURSING = 'NURSING',
    PHARMACY = 'PHARMACY',
    VETERINARY = 'VETERINARY',
    SOCIAL_WORK = 'SOCIAL_WORK',
    BIOMEDICINE = 'BIOMEDICINE',
    SPEECH_THERAPY = 'SPEECH_THERAPY',
    OCCUPATIONAL_THERAPY = 'OCCUPATIONAL_THERAPY',
    NUTRITION = 'NUTRITION',
    PHYSICAL_EDUCATION = 'PHYSICAL_EDUCATION',      
    RADIOLOGY = 'RADIOLOGY',                        
    OPTOMETRY = 'OPTOMETRY',                        
    SANITARY = 'SANITARY',                         
    NURSING_TECHNICIAN = 'NURSING_TECHNICIAN'
}